+++ resources/views/components/form/course-type-radio.blade.php
@@ -6,7 +6,7 @@
         <div class="bg-white rounded-md -space-y-px shadow-sm" x-data="{ type:''}">
             <label @click="type = 'class'" :class="{'bg-indigo-50 border-indigo-200 z-10': type == 'class' }"
                 class="border-gray-200 rounded-tl-md rounded-tr-md relative border p-4 flex cursor-pointer">
-                <input type="radio" name="type" value="Class" {{ $attributes }}
+                <input type="radio" name="type" value="class" {{ $attributes }}
                     class="h-4 w-4 mt-0.5 cursor-pointer text-indigo-600 border-gray-300 focus:ring-indigo-500">
                 <div class="ml-3 flex flex-col">
                     <span :class="{'text-indigo-900': type === 'class' }"
@@ -23,7 +23,7 @@
 
             <label @click="type = 'workshop'" :class="{'bg-indigo-50 border-indigo-200 z-10': type === 'workshop' }"
                 class="border-gray-200 relative border p-4 flex cursor-pointer">
-                <input type="radio" name="type" value="Workshop" {{ $attributes }}
+                <input type="radio" name="type" value="workshop" {{ $attributes }}
                     class="h-4 w-4 mt-0.5 cursor-pointer text-indigo-600 border-gray-300 focus:ring-indigo-500">
                 <div class="ml-3 flex flex-col">
                     <span id="privacy-setting-1-label" :class="{'text-indigo-900': type === 'workshop' }"
@@ -40,7 +40,7 @@
             <!-- Checked: "bg-indigo-50 border-indigo-200 z-10", Not Checked: "border-gray-200" -->
             <label @click="type = 'bootcamp'" :class="{'bg-indigo-50 border-indigo-200 z-10': type === 'bootcamp' }"
                 class="border-gray-200 rounded-bl-md rounded-br-md relative border p-4 flex cursor-pointer">
-                <input type="radio" name="type" value="Bootcamp" {{ $attributes }}
+                <input type="radio" name="type" value="bootcamp" {{ $attributes }}
                     class="h-4 w-4 mt-0.5 cursor-pointer text-indigo-600 border-gray-300 focus:ring-indigo-500">
                 <div class="ml-3 flex flex-col">
                     <span id="privacy-setting-2-label" :class="{'text-indigo-900': type === 'bootcamp' }"