+++ resources/views/course/dashboard.blade.php
@@ -1,16 +1,14 @@
 <x-app-layout>
     <x-slot name="header">
-        <div class="md:flex md:items-center md:justify-between">
+        <div class="flex items-center justify-between">
             <div class="flex-1 min-w-0">
                 <x-typo.page-heading title="{{ $course->name}}" />
             </div>
-            <div class="mt-4 flex md:mt-0 md:ml-4">
-
+            <div class="flex md:mt-0 md:ml-4">
                 <a href="{{ url()->previous() }}"
                     class="inline-flex items-center px-4 py-2 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
                     Back
                 </a>
-
                 @can('update', $course)
                 <a href="{{ route('course.edit', $course) }}"
                     class="ml-3 inline-flex items-center px-4 py-2 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
@@ -28,6 +26,7 @@
                     @include('course.dashboard.left')
                 </div>
                 <main class="lg:col-span-9 xl:col-span-6">
+                    @include('course.dashboard._mobile-menu')
                     @include('course.dashboard.main')
                 </main>
                 <aside class="hidden xl:block xl:col-span-4">