+++ resources/views/course/dashboard/students.blade.php
@@ -1,3 +1,4 @@
+@include('course.dashboard._mobile-menu')
 <div class="my-4">
     <livewire:shared.registered-table />
 </div>